Se unieron los endpoints que estaban en archivos separados ademas de agregarle el middleware Api
