+++ app.js
@@ -6,17 +6,14 @@ const dbconnect = require("./config/conexion");
 const cors = require('cors')
 const rutas = require('./routes/login')
 require('dotenv').config();
-const platillosRoutes = require('./routes/comida_router');
-
-//const comidaRouter= require('./routes/comidaRouter')
+const comidaRouter= require('./routes/comida_router')
 
 app.use(express.urlencoded({extended:false}))
 app.use(express.json());
 app.use(cors());
 
 app.use(morgan('dev'))
-//app.use('/Api', comidaRouter)
-app.use(platillosRoutes);
+app.use('/Api', comidaRouter)
 app.use(rutas)
 
 app.get("/", (req,res)=>{
